Solver classes renamed
